Skeleton dev.sites.emr.ge

+++ php-classes/Tag.class.php
@@ -14,9 +14,12 @@ class Tag extends ActiveRecord
     static public $collectionRoute = '/tags';
 
     public static $fields = array(
-        'Title'
+        'Title' => array(
+            'includeInSummary' => true
+        )
         ,'Handle' => array(
-            'unique' => true
+            'unique' => true,
+            'includeInSummary' => true
         )
         ,'Description'
     );
@@ -88,42 +91,46 @@ class Tag extends ActiveRecord
             $tags = $newTags;
         }
 
-        foreach ($tags AS $tagTitle) {
-            if (!$tagTitle) {
+        foreach ($tags AS $tag) {
+            if ($tag && (is_string($tag) || is_int($tag))) {
+                $tag = static::getFromHandle($tag, $autoCreate, $prefix);
+            }
+
+            if (!$tag) {
                 continue;
             }
 
-            if ($Tag = static::getFromHandle($tagTitle, $autoCreate)) {
-                $Tag->assignItem($Context);
-                $assignedTags[$Tag->ID] = $Tag;
-            }
+            $tag->assignItem($Context);
+            $assignedTags[$tag->ID] = $tag;
         }
 
         if ($prefix) {
-            $groupTags = DB::allValues('ID', 'SELECT ID FROM `%s` WHERE Handle LIKE "%s.%%"', array(
+            $prefixTags = DB::allValues('ID', 'SELECT ID FROM `%s` WHERE Handle LIKE "%s.%%"', array(
                 \Tag::$tableName,
                 DB::escape($prefix)
             ));
         }
 
         // delete tags
-        try {
-            DB::query(
-                'DELETE FROM `%s` WHERE ContextClass = "%s" AND ContextID = %u AND TagID NOT IN (%s) %s'
-                ,array(
-                    TagItem::$tableName
-                    ,DB::escape($Context->getRootClass())
-                    ,$Context->ID
-                    ,count($assignedTags) ? join(',', array_keys($assignedTags)) : '0'
-                    ,$prefix ? (' AND TagID IN ('.implode(',', $groupTags).')') : ''
-                )
-            );
-        } catch (TableNotFoundException $e) { }
+        if (!$prefix || count($prefixTags)) { // if there are no existing tags with this prefix there are no items to delete
+            try {
+                DB::query(
+                    'DELETE FROM `%s` WHERE ContextClass = "%s" AND ContextID = %u AND TagID NOT IN (%s) %s'
+                    ,array(
+                        TagItem::$tableName
+                        ,DB::escape($Context->getRootClass())
+                        ,$Context->ID
+                        ,count($assignedTags) ? join(',', array_keys($assignedTags)) : '0'
+                        ,!empty($prefixTags) ? (' AND TagID IN ('.implode(',', $prefixTags).')') : ''
+                    )
+                );
+            } catch (TableNotFoundException $e) { }
+        }
 
         return $assignedTags;
     }
 
-    public static function getFromHandle($handle, $autoCreate = true)
+    public static function getFromHandle($handle, $autoCreate = true, $prefix = null)
     {
         $Tag = false;
 
@@ -136,11 +143,14 @@ class Tag extends ActiveRecord
         }
 
         if (!$Tag) {
-            $Tag = Tag::getByTitle($handle);
+            $Tag = Tag::getByTitle($handle, $prefix);
         }
 
         if (!$Tag && $autoCreate) {
-            $Tag = Tag::create(array('Title' => $handle), true);
+            $Tag = Tag::create(array(
+                'Title' => $handle,
+                'Handle' => HandleBehavior::getUniqueHandle(__CLASS__, $prefix ? "$prefix.$handle" : $handle)
+            ), true);
         }
 
         return $Tag;
@@ -158,9 +168,16 @@ class Tag extends ActiveRecord
         ));
     }
 
-    public static function getByTitle($title)
+    public static function getByTitle($title, $prefix = null)
     {
-        return static::getByField('Title', $title, true);
+        if ($prefix) {
+            return static::getByWhere(array(
+                'Title' => $title,
+                'Handle LIKE "'.DB::escape($prefix).'.%%"'
+            ));
+        } else {
+            return static::getByField('Title', $title, true);
+        }
     }
 
     public function getValue($name)
@@ -366,10 +383,16 @@ class Tag extends ActiveRecord
         return $class::instantiateRecords(DB::allRecords($classQuery));
     }
 
-    public static function getTagsString($tags)
+    public static function getTagsString($tags, $prefix = null)
     {
+        if ($prefix) {
+            $tags = array_filter($tags, function($Tag) use ($prefix) {
+                return $Tag->HandlePrefix == $prefix;
+            });
+        }
+
         return implode(', ', array_map(function($Tag) {
-            return $Tag->Handle;
+            return $Tag->Title;
         }, $tags));
     }
